Use https with the `wsk` CLI on minishift too
The `wsk` CLI really expects the api host to be https, and in at least
some places prepends the string 'https://' to its value. So, switch
the minishift instructions to use https just like the minikube
instructions already did. This also removes the unnecessary 'https://'
prefix from the minikube instructions.